support more dtypes

+++ dimble/dimble.py
@@ -8,7 +8,7 @@ from safetensors.numpy import save_file
 from dimble_rs import dimble_rs
 
 
-def _dicom_to_ir(dicom_path: Path, output_name: str) -> dict[str, Path]:
+def _dicom_to_ir(dicom_path: Path, output_name: str, dtype=np.float32) -> dict[str, Path]:
     ds = pydicom.dcmread(dicom_path)
     output_json = Path(output_name + ".json")
     output_pixel_array = Path(output_name + ".safetensors")
@@ -16,7 +16,7 @@ def _dicom_to_ir(dicom_path: Path, output_name: str) -> dict[str, Path]:
         ds_json_dict = ds.to_json_dict()
         json.dump(ds_json_dict, f, sort_keys=True, indent=4)
 
-    pixel_array = ds.pixel_array.astype(np.float32)
+    pixel_array = ds.pixel_array.astype(dtype)
     save_file({"pixel_array": pixel_array}, output_pixel_array)
     return {"json": output_json, "pixel_array": output_pixel_array}
 
@@ -29,9 +29,9 @@ def _dimble_to_ir(dimble_path: Path, output_path: Path) -> None:
     dimble_rs.dimble_to_dicom_json(str(dimble_path), str(output_path))
 
 
-def dicom_to_dimble(dicom_path: Path, output_path: Path) -> None:
+def dicom_to_dimble(dicom_path: Path, output_path: Path, dtype=np.float32) -> None:
     dicom_path = Path(dicom_path)
-    ir_paths = _dicom_to_ir(dicom_path, str(Path("/tmp") / (dicom_path.stem + ".ir")))
+    ir_paths = _dicom_to_ir(dicom_path, str(Path("/tmp") / (dicom_path.stem + ".ir")), dtype=dtype)
     try:
         _ir_to_dimble(ir_paths["json"], ir_paths["pixel_array"], output_path)
     finally:
